Added rcon (remote control) tool and exiting nicely feature.
gcb rcon allows you to send commands to gcb through sockets. There is a basic password authentication. Protocol is text-based and you can use telnet as the rcon client. Currently only "exit" and "exit nicely" are commands.

Exiting nicely allows gcb to exit without closing currently open virtual TCP connections. Note that further modifications are required for gcb to be able to restart after exiting nicely, since the UDP port is not closed until all virtual TCP connections are closed (otherwise they would be disconnected). The plan is to allow an auto feature to let gcb select a peer port randomly, and re-select if the chosen port is in use.


git-svn-id: http://gcb.googlecode.com/svn/trunk@217 35299b14-8a3f-7d2e-6494-bf2248573a8d